+++ src/stories/MastheadPrimary.stories.js
@@ -5,7 +5,32 @@ export default {
     title: "MASTHEAD / Primary",
     component: MastheadPrimary,
 }
+const mock = {
+    linkItems: [
+        {
+            text: "Course Reserves",
+            url: "https://catalog.library.ucla.edu/vwebv/enterCourseReserve.do",
+        },
+        {
+            text: "UCLA Research Guides",
+            url: "https://guides.library.ucla.edu/",
+        },
+        {
+            text: "Databases A-Z",
+            url: "https://guides.library.ucla.edu/az.php",
+        },
+    ],
+    advancedSearchLink: {
+        text: "Advanced Search",
+        url: "https://www.library.ucla.edu/search",
+    },
+}
 export const Default = () => ({
-    template: `<masthead-primary/>`,
+    data() {
+        return {
+            ...mock,
+        }
+    },
+    template: `<masthead-primary :link-items="linkItems" :advanced-search-link="advancedSearchLink"/>`,
     components: { MastheadPrimary },
 })