+++ CHANGELOG.md
@@ -1,3 +1,45 @@
+## [2.11.2](https://github.com/UCLALibrary/ucla-library-website-components/compare/v2.11.1...v2.11.2) (2022-12-05)
+
+
+### Bug Fixes
+
+* bannerFeatured - choose to link to title ([#267](https://github.com/UCLALibrary/ucla-library-website-components/issues/267)) ([0d2b283](https://github.com/UCLALibrary/ucla-library-website-components/commit/0d2b2831b7198b7f2f2a39f1eaa32cf97466f6e9))
+
+## [2.11.1](https://github.com/UCLALibrary/ucla-library-website-components/compare/v2.11.0...v2.11.1) (2022-12-03)
+
+
+### Bug Fixes
+
+* update masthead primary story and add smart-link to search home ([#268](https://github.com/UCLALibrary/ucla-library-website-components/issues/268)) ([ba427c9](https://github.com/UCLALibrary/ucla-library-website-components/commit/ba427c9e72e5736e461b23231bfbaee00af0654e))
+
+# [2.11.0](https://github.com/UCLALibrary/ucla-library-website-components/compare/v2.10.0...v2.11.0) (2022-12-02)
+
+
+### Features
+
+* APPS-2043 add checkbox component ([#259](https://github.com/UCLALibrary/ucla-library-website-components/issues/259)) ([2e3dbc4](https://github.com/UCLALibrary/ucla-library-website-components/commit/2e3dbc44c959ab22237608b9518a7fbb5f4d87df))
+
+# [2.10.0](https://github.com/UCLALibrary/ucla-library-website-components/compare/v2.9.5...v2.10.0) (2022-12-02)
+
+
+### Features
+
+* updates media gallery, impact carousel, and card spacing for collections ([#265](https://github.com/UCLALibrary/ucla-library-website-components/issues/265)) ([441ee51](https://github.com/UCLALibrary/ucla-library-website-components/commit/441ee5103dc2ff9b2e505b3a2bfc8c2b35be81bd))
+
+## [2.9.5](https://github.com/UCLALibrary/ucla-library-website-components/compare/v2.9.4...v2.9.5) (2022-12-02)
+
+
+### Bug Fixes
+
+* APPS-2080 Link title on the bannerFeatured component ([#264](https://github.com/UCLALibrary/ucla-library-website-components/issues/264)) ([7554f7e](https://github.com/UCLALibrary/ucla-library-website-components/commit/7554f7e06faf9753bd23bbdee76d95c18114cc40))
+
+## [2.9.4](https://github.com/UCLALibrary/ucla-library-website-components/compare/v2.9.3...v2.9.4) (2022-11-30)
+
+
+### Bug Fixes
+
+* Update Amenities on LocationLists ([#262](https://github.com/UCLALibrary/ucla-library-website-components/issues/262)) ([eed2249](https://github.com/UCLALibrary/ucla-library-website-components/commit/eed22495628f59d1045dc00d713be60e4e060cf3))
+
 ## [2.9.3](https://github.com/UCLALibrary/ucla-library-website-components/compare/v2.9.2...v2.9.3) (2022-11-30)